requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.5
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
